backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
google-adk
litellm
ddgs
